+++ packages/alloy_generator/lib/src/emitters/root_scope_emitter.dart
@@ -100,9 +100,8 @@ class RootScopeEmitter {
         ? factory.property('create').call([refer('scope')])
         : factory;
 
-    return refer('scope')
-        .property(method)
-        .call([argument], named, [exposed])
-        .statement;
+    return refer(
+      'scope',
+    ).property(method).call([argument], named, [exposed]).statement;
   }
 }